Refactor to share logic and remove circular imports
