+++ routemaster/state_machine/__init__.py
@@ -7,10 +7,10 @@ from routemaster.state_machine.api import (
     delete_label,
     get_label_state,
     get_label_metadata,
+    process_action_retries,
     update_metadata_for_label,
 )
-from routemaster.state_machine.actions import \
-    process_retries as process_action_retries
+from routemaster.state_machine.types import CronProcessor, IsExitingCheck
 from routemaster.state_machine.exceptions import (
     UnknownLabel,
     LabelAlreadyExists,
@@ -23,6 +23,8 @@ __all__ = (
     'create_label',
     'delete_label',
     'UnknownLabel',
+    'CronProcessor',
+    'IsExitingCheck',
     'get_label_state',
     'get_label_metadata',
     'LabelAlreadyExists',
