+++ routemaster/state_machine/tests/test_actions.py
@@ -2,7 +2,8 @@ import mock
 import pytest
 
 from routemaster.webhooks import WebhookResult
-from routemaster.state_machine.actions import process_action, process_retries
+from routemaster.state_machine.api import process_action_retries
+from routemaster.state_machine.actions import process_action
 from routemaster.state_machine.exceptions import DeletedLabel
 
 
@@ -16,7 +17,7 @@ def test_actions_are_run_and_states_advanced(app_config, create_label, mock_webh
 
     # Now attempt to process the retry of the action, succeeding.
     with mock_webhook(WebhookResult.SUCCESS) as webhook:
-        process_retries(
+        process_action_retries(
             app_config,
             state_machine,
             state_machine.states[1],
@@ -45,7 +46,7 @@ def test_actions_do_not_advance_state_on_fail(app_config, create_label, mock_web
         create_label('foo', state_machine.name, {'should_progress': True})
 
     with mock_webhook(WebhookResult.FAIL) as webhook:
-        process_retries(
+        process_action_retries(
             app_config,
             state_machine,
             state_machine.states[1],
@@ -79,9 +80,9 @@ def test_action_retry_trigger_continues_as_far_as_possible(app_config, create_la
     # Now attempt to process the retry of the action, succeeding.
     with mock_webhook(WebhookResult.SUCCESS) as webhook:
         with mock.patch(
-            'routemaster.state_machine.actions.process_transitions',
+            'routemaster.state_machine.api.process_transitions',
         ) as mock_process_transitions:
-            process_retries(
+            process_action_retries(
                 app_config,
                 state_machine,
                 state_machine.states[1],
@@ -211,7 +212,7 @@ def test_process_action_fails_retry_works(app_config, create_label, mock_webhook
 
     # Now retry with succeeding webhook
     with mock_webhook(WebhookResult.SUCCESS) as webhook:
-        process_retries(app_config, state_machine, action, lambda: False)
+        process_action_retries(app_config, state_machine, action, lambda: False)
         webhook.assert_called_once()
 
     assert_history([
